refactor: remove superflous cleanup
The transaction removal found in the `check_for_rejection_with_caching`
is actually unnecessary, because the callers either don't add anything
to the mempool, or perform the cleanup a bit later.
Otherwise, it is benefitial to remove the access to the mempool from
this function, so when we have locks around the mempool, that doesn't
have to be acquired.

Signed-off-by: Adam Mozes <adam.mozes@rdx.works>